or1k: sysdep: dont clobber syscall regs when we have memcpy
This fixes a issue with posix_spawn which uses rt_procmask sycall.
The mask is about 0x80 bytes and setting the function arg requires
a builtin_memcpy.  That memcpy is carried out during argument setting.
When the arguments are required to be in registers they get clobbered.

It seems a similar fix was made in mips.  The fix is to first store
the arguments to long registers then to register long registers to
avoid the register clobbering.

This was pointed out by support/test_capture_process.